styling + namng fixes
